requirements.txt
discord.py>=2.3.0
aiosqlite>=0.19.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
psutil>=5.9.0
yt-dlp>=2024.1.0
PyNaCl>=1.5.0

Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0
soundfile>=0.12.0
discord-ext-voice_recv>=0.5.0
